+++ commons/src/main/java/de/gematik/refv/commons/configuration/ValidationMessageTransformation.java
@@ -25,4 +25,5 @@ public class ValidationMessageTransformation {
     String severityLevelFrom;
     String severityLevelTo;
     String locatorString;
+    String messageId;
 }