+++ commons/src/main/java/de/gematik/refv/commons/validation/ValidationModule.java
@@ -33,5 +33,5 @@ public interface ValidationModule {
     ValidationResult validateFile(@NonNull String inputPath) throws IllegalArgumentException, IOException;
     ValidationResult validateString(@NonNull String fhirResourceAsString) throws IllegalArgumentException;
     ValidationResult validateFile(Path inputPath) throws IllegalArgumentException, IOException;
-    void initialize() throws ValidationModuleInitializationException, IOException;
+    void initialize() throws ValidationModuleInitializationException;
 }